# Reduce invalidation by not specializing lastindex/axes1