@jishnub 2022-10-09
We may reduce another set of invalidation by not specializing `Base.axes1` for `IdOffsetRange`, which seems redundant as well on recent Julia versions.